Walk the reflog to print out the recent history for branch changes
It only outputs branches that still exist so merged branches won't show
up and it lists the most recent last so it is easiest to find at the end
of the command output. It also removes duplicates so if you switch
between branches a bunch that branch will not show up multiple times.