Fix git history to prevent an infinite loop
If a rename happened and then the branch was renamed back to the
original name, an infinite loop would be triggered because it would
infinitely try and look up the original name.

Instead, it now deletes an entry from the rename lookup if it is the
target of a rename because once the rename happens, the chain is broken
since the branch is no longer related to the old branch.