Add a few more utils
The `git-main` util is a way of getting the main branch name without
relying on master or main as the default. It still assumes that it's one
of those two names but will just generally check which of the two
branches exist.

The `git-fixup` attribute will automatically find the first commit in a
chain in for a PR and will invoke `git commit --fixup` using that
commit. It will also add `--edit` to ensure the editor comes up by
default. This can be overridden by just using `--no-edit` as a
parameter.